feat: listagem de consultas totais prontas

+++ backend-dotnet/backend-dotnet/DTO/Person/ListAllPersonsDTO.cs
@@ -7,7 +7,7 @@ public class ListAllPersonsDTO
     public Guid Id { get; set; }
     public string Name { get; set; }
     public int Age { get; set; }
-    public int TotalIncome { get; set; }
-    public int TotalExpenses { get; set; }
-    public int Balance { get; set; }
+    public decimal TotalIncome { get; set; }
+    public decimal TotalExpenses { get; set; }
+    public decimal Balance { get; set; }
 }
